refactor(web-ai-api-check): split modal host UI components
