+++ src/entrypoints/content/webAiApiCheck/components/apiCheckModalTypes.ts
@@ -1,0 +1,12 @@
+import type {
+  ApiVerificationProbeId,
+  ApiVerificationProbeResult,
+} from "~/services/verification/aiApiVerification"
+
+export type ProbeItemState = {
+  id: ApiVerificationProbeId
+  requiresModelId: boolean
+  isRunning: boolean
+  attempts: number
+  result: ApiVerificationProbeResult | null
+}
